+++ alacritty/src/event.rs
@@ -1232,10 +1232,12 @@ impl<'a, N: Notify + 'a, T: EventListener> input::ActionContext<T> for ActionCon
         }
 
         let hint_bounds = hint.bounds();
-        let text = match hint.text(self.terminal) {
+        let raw_text = match hint.text(self.terminal) {
             Some(text) => text,
             None => return,
         };
+
+        let text = hint.substitute_text(raw_text);
 
         match &hint.action() {
             // Launch an external program.